@@ -3,7 +3,15 @@ package core
 import (
 	"bytes"
 	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+	"sort"
 	"strconv"
+	"strings"
+
+	"github.com/danielmiessler/fabric/plugins/ai/exolab"
+	"github.com/danielmiessler/fabric/plugins/strategy"
 
 	"github.com/samber/lo"
 
@@ -14,20 +22,19 @@ import (
 	"github.com/danielmiessler/fabric/plugins/ai/azure"
 	"github.com/danielmiessler/fabric/plugins/ai/dryrun"
 	"github.com/danielmiessler/fabric/plugins/ai/gemini"
-	"github.com/danielmiessler/fabric/plugins/ai/groq"
-	"github.com/danielmiessler/fabric/plugins/ai/mistral"
+	"github.com/danielmiessler/fabric/plugins/ai/lmstudio"
 	"github.com/danielmiessler/fabric/plugins/ai/ollama"
 	"github.com/danielmiessler/fabric/plugins/ai/openai"
-	"github.com/danielmiessler/fabric/plugins/ai/openrouter"
-	"github.com/danielmiessler/fabric/plugins/ai/siliconcloud"
+	"github.com/danielmiessler/fabric/plugins/ai/openai_compatible"
 	"github.com/danielmiessler/fabric/plugins/db/fsdb"
+	"github.com/danielmiessler/fabric/plugins/template"
 	"github.com/danielmiessler/fabric/plugins/tools"
 	"github.com/danielmiessler/fabric/plugins/tools/jina"
 	"github.com/danielmiessler/fabric/plugins/tools/lang"
 	"github.com/danielmiessler/fabric/plugins/tools/youtube"
 )
 
-func NewPluginRegistry(db *fsdb.Db) (ret *PluginRegistry) {
+func NewPluginRegistry(db *fsdb.Db) (ret *PluginRegistry, err error) {
 	ret = &PluginRegistry{
 		Db:             db,
 		VendorManager:  ai.NewVendorsManager(),
@@ -36,27 +43,72 @@ func NewPluginRegistry(db *fsdb.Db) (ret *PluginRegistry) {
 		YouTube:        youtube.NewYouTube(),
 		Language:       lang.NewLanguage(),
 		Jina:           jina.NewClient(),
+		Strategies:     strategy.NewStrategiesManager(),
 	}
 
-	ret.Defaults = tools.NeeDefaults(ret.VendorManager.GetModels)
+	var homedir string
+	if homedir, err = os.UserHomeDir(); err != nil {
+		return
+	}
+	ret.TemplateExtensions = template.NewExtensionManager(filepath.Join(homedir, ".config/fabric"))
 
-	ret.VendorsAll.AddVendors(openai.NewClient(), ollama.NewClient(), azure.NewClient(), groq.NewClient(),
-		gemini.NewClient(), anthropic.NewClient(), siliconcloud.NewClient(), openrouter.NewClient(), mistral.NewClient())
+	ret.Defaults = tools.NeeDefaults(ret.GetModels)
+
+	// Create a vendors slice to hold all vendors (order doesn't matter initially)
+	vendors := []ai.Vendor{}
+
+	// Add non-OpenAI compatible clients
+	vendors = append(vendors,
+		openai.NewClient(),
+		ollama.NewClient(),
+		azure.NewClient(),
+		gemini.NewClient(),
+		anthropic.NewClient(),
+		lmstudio.NewClient(),
+		exolab.NewClient(),
+	)
+
+	// Add all OpenAI-compatible providers
+	for providerName := range openai_compatible.ProviderMap {
+		provider, _ := openai_compatible.GetProviderByName(providerName)
+		vendors = append(vendors, openai_compatible.NewClient(provider))
+	}
+
+	// Sort vendors by name for consistent ordering (case-insensitive)
+	sort.Slice(vendors, func(i, j int) bool {
+		return strings.ToLower(vendors[i].GetName()) < strings.ToLower(vendors[j].GetName())
+	})
+
+	// Add all sorted vendors to VendorsAll
+	ret.VendorsAll.AddVendors(vendors...)
 	_ = ret.Configure()
 
 	return
 }
 
+func (o *PluginRegistry) ListVendors(out io.Writer) error {
+	vendors := lo.Map(o.VendorsAll.Vendors, func(vendor ai.Vendor, _ int) string {
+		return vendor.GetName()
+	})
+	fmt.Fprint(out, "Available Vendors:\n\n")
+	for _, vendor := range vendors {
+		fmt.Fprintf(out, "%s\n", vendor)
+	}
+	return nil
+}
+
 type PluginRegistry struct {
 	Db *fsdb.Db
 
-	VendorManager  *ai.VendorsManager
-	VendorsAll     *ai.VendorsManager
-	Defaults       *tools.Defaults
-	PatternsLoader *tools.PatternsLoader
-	YouTube        *youtube.YouTube
-	Language       *lang.Language
-	Jina           *jina.Client
+	VendorManager      *ai.VendorsManager
+	VendorsAll         *ai.VendorsManager
+	Defaults           *tools.Defaults
+	PatternsLoader     *tools.PatternsLoader
+	YouTube            *youtube.YouTube
+	Language           *lang.Language
+	Jina               *jina.Client
+	TemplateExtensions *template.ExtensionManager
+	Strategies         *strategy.StrategiesManager
 }
 
 func (o *PluginRegistry) SaveEnvFile() (err error) {
@@ -65,6 +117,7 @@ func (o *PluginRegistry) SaveEnvFile() (err error) {
 
 	o.Defaults.Settings.FillEnvFileContent(&envFileContent)
 	o.PatternsLoader.SetupFillEnvFileContent(&envFileContent)
+	o.Strategies.SetupFillEnvFileContent(&envFileContent)
 
 	for _, vendor := range o.VendorManager.Vendors {
 		vendor.SetupFillEnvFileContent(&envFileContent)
@@ -80,7 +133,7 @@ func (o *PluginRegistry) SaveEnvFile() (err error) {
 
 func (o *PluginRegistry) Setup() (err error) {
 	setupQuestion := plugins.NewSetupQuestion("Enter the number of the plugin to setup")
-	groupsPlugins := common.NewGroupsItemsSelector[plugins.Plugin]("Available plugins",
+	groupsPlugins := common.NewGroupsItemsSelector("Available plugins (please configure all required plugins):",
 		func(plugin plugins.Plugin) string {
 			var configuredLabel string
 			if plugin.IsConfigured() {
@@ -96,10 +149,10 @@ func (o *PluginRegistry) Setup() (err error) {
 			return vendor
 		})...)
 
-	groupsPlugins.AddGroupItems("Tools", o.Defaults, o.PatternsLoader, o.YouTube, o.Language, o.Jina)
+	groupsPlugins.AddGroupItems("Tools", o.Defaults, o.Jina, o.Language, o.PatternsLoader, o.Strategies, o.YouTube)
 
 	for {
-		groupsPlugins.Print()
+		groupsPlugins.Print(false)
 
 		if answerErr := setupQuestion.Ask("Plugin Number"); answerErr != nil {
 			break
@@ -126,7 +179,8 @@ func (o *PluginRegistry) Setup() (err error) {
 			}
 
 			if _, ok := o.VendorManager.VendorsByName[plugin.GetName()]; !ok {
-				if vendor, ok := plugin.(ai.Vendor); ok {
+				var vendor ai.Vendor
+				if vendor, ok = plugin.(ai.Vendor); ok {
 					o.VendorManager.AddVendors(vendor)
 				}
 			}
@@ -148,13 +202,24 @@ func (o *PluginRegistry) SetupVendor(vendorName string) (err error) {
 	return
 }
 
-// Configure buildClient VendorsController based on the environment variables
-func (o *PluginRegistry) Configure() (err error) {
+func (o *PluginRegistry) ConfigureVendors() {
+	o.VendorManager.Clear()
 	for _, vendor := range o.VendorsAll.Vendors {
 		if vendorErr := vendor.Configure(); vendorErr == nil {
 			o.VendorManager.AddVendors(vendor)
 		}
 	}
+}
+
+func (o *PluginRegistry) GetModels() (ret *ai.VendorsModels, err error) {
+	o.ConfigureVendors()
+	ret, err = o.VendorManager.GetModels()
+	return
+}
+
+// Configure buildClient VendorsController based on the environment variables
+func (o *PluginRegistry) Configure() (err error) {
+	o.ConfigureVendors()
 	_ = o.Defaults.Configure()
 	_ = o.PatternsLoader.Configure()
 
@@ -165,7 +230,7 @@ func (o *PluginRegistry) Configure() (err error) {
 	return
 }
 
-func (o *PluginRegistry) GetChatter(model string, modelContextLength int, stream bool, dryRun bool) (ret *Chatter, err error) {
+func (o *PluginRegistry) GetChatter(model string, modelContextLength int, strategy string, stream bool, dryRun bool) (ret *Chatter, err error) {
 	ret = &Chatter{
 		db:     o.Db,
 		Stream: stream,
@@ -206,10 +271,17 @@ func (o *PluginRegistry) GetChatter(model string, modelContextLength int, stream
 	}
 
 	if ret.vendor == nil {
+		var errMsg string
+		if defaultModel == "" || defaultVendor == "" {
+			errMsg = "Please run, fabric --setup, and select default model and vendor."
+		} else {
+			errMsg = "could not find vendor."
+		}
 		err = fmt.Errorf(
-			"could not find vendor.\n Model = %s\n Model = %s\n Vendor = %s",
-			model, defaultModel, defaultVendor)
+			" Requested Model = %s\n Default Model = %s\n Default Vendor = %s.\n\n%s",
+			model, defaultModel, defaultVendor, errMsg)
 		return
 	}
+	ret.strategy = strategy
 	return
 }
