@@ -1,10 +1,10 @@
 package azure
 
 import (
-	"github.com/danielmiessler/fabric/plugins"
-	"github.com/danielmiessler/fabric/plugins/ai/openai"
 	"strings"
 
+	"github.com/danielmiessler/fabric/plugins"
+	"github.com/danielmiessler/fabric/plugins/ai/openai"
 	goopenai "github.com/sashabaranov/go-openai"
 )
 
@@ -13,6 +13,8 @@ func NewClient() (ret *Client) {
 	ret.Client = openai.NewClientCompatible("Azure", "", ret.configure)
 	ret.ApiDeployments = ret.AddSetupQuestionCustom("deployments", true,
 		"Enter your Azure deployments (comma separated)")
+	ret.ApiVersion = ret.AddSetupQuestionCustom("API Version", false,
+		"Enter the Azure API version (optional)")
 
 	return
 }
@@ -20,19 +22,26 @@ func NewClient() (ret *Client) {
 type Client struct {
 	*openai.Client
 	ApiDeployments *plugins.SetupQuestion
+	ApiVersion     *plugins.SetupQuestion
 
 	apiDeployments []string
 }
 
 func (oi *Client) configure() (err error) {
 	oi.apiDeployments = strings.Split(oi.ApiDeployments.Value, ",")
-	oi.ApiClient = goopenai.NewClientWithConfig(goopenai.DefaultAzureConfig(oi.ApiKey.Value, oi.ApiBaseURL.Value))
+	config := goopenai.DefaultAzureConfig(oi.ApiKey.Value, oi.ApiBaseURL.Value)
+	if oi.ApiVersion.Value != "" {
+		config.APIVersion = oi.ApiVersion.Value
+	}
+	oi.ApiClient = goopenai.NewClientWithConfig(config)
 	return
 }
 
 func (oi *Client) ListModels() (ret []string, err error) {
-	// Azure Open Client has models and deployments. We need to get the deployments to use them for chat
-	// There is no easy way to get the deployments from the API, so we need to ask the user to provide them
 	ret = oi.apiDeployments
 	return
+}
+
+func (oi *Client) NeedsRawMode(modelName string) bool {
+	return false
 }
