@@ -4,8 +4,11 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"github.com/danielmiessler/fabric/plugins"
+	"sort"
+	"strings"
 	"sync"
+
+	"github.com/danielmiessler/fabric/plugins"
 )
 
 func NewVendorsManager() *VendorsManager {
@@ -27,6 +30,12 @@ func (o *VendorsManager) AddVendors(vendors ...Vendor) {
 		o.VendorsByName[vendor.GetName()] = vendor
 		o.Vendors = append(o.Vendors, vendor)
 	}
+}
+
+func (o *VendorsManager) Clear(vendors ...Vendor) {
+	o.VendorsByName = map[string]Vendor{}
+	o.Vendors = []Vendor{}
+	o.Models = nil
 }
 
 func (o *VendorsManager) SetupFillEnvFileContent(envFileContent *bytes.Buffer) {
@@ -87,8 +96,10 @@ func (o *VendorsManager) readModels() (err error) {
 	for result := range resultsChan {
 		if result.err != nil {
 			fmt.Println(result.vendorName, result.err)
-			cancel() // Cancel remaining goroutines if needed
 		} else {
+			sort.Slice(result.models, func(i, j int) bool {
+				return strings.ToLower(result.models[i]) < strings.ToLower(result.models[j])
+			})
 			o.Models.AddGroupItems(result.vendorName, result.models...)
 		}
 	}
