@@ -5,7 +5,12 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/danielmiessler/fabric/common"
+	"github.com/danielmiessler/fabric/plugins/template"
 )
+
+const inputSentinel = "__FABRIC_INPUT_SENTINEL_TOKEN__"
 
 type PatternsEntity struct {
 	*StorageEntity
@@ -13,7 +18,77 @@ type PatternsEntity struct {
 	UniquePatternsFilePath string
 }
 
-func (o *PatternsEntity) Get(name string) (ret *Pattern, err error) {
+// Pattern represents a single pattern with its metadata
+type Pattern struct {
+	Name        string
+	Description string
+	Pattern     string
+}
+
+// GetApplyVariables main entry point for getting patterns from any source
+func (o *PatternsEntity) GetApplyVariables(
+	source string, variables map[string]string, input string) (pattern *Pattern, err error) {
+
+	// Determine if this is a file path
+	isFilePath := strings.HasPrefix(source, "\\") ||
+		strings.HasPrefix(source, "/") ||
+		strings.HasPrefix(source, "~") ||
+		strings.HasPrefix(source, ".")
+
+	if isFilePath {
+		// Resolve the file path using GetAbsolutePath
+		absPath, err := common.GetAbsolutePath(source)
+		if err != nil {
+			return nil, fmt.Errorf("could not resolve file path: %v", err)
+		}
+
+		// Use the resolved absolute path to get the pattern
+		pattern, err = o.getFromFile(absPath)
+	} else {
+		// Otherwise, get the pattern from the database
+		pattern, err = o.getFromDB(source)
+	}
+
+	if err != nil {
+		return
+	}
+
+	// Apply variables to the pattern
+	err = o.applyVariables(pattern, variables, input)
+	return
+}
+
+func (o *PatternsEntity) applyVariables(
+	pattern *Pattern, variables map[string]string, input string) (err error) {
+
+	// Ensure pattern has an {{input}} placeholder
+	// If not present, append it on a new line
+	if !strings.Contains(pattern.Pattern, "{{input}}") {
+		if !strings.HasSuffix(pattern.Pattern, "\n") {
+			pattern.Pattern += "\n"
+		}
+		pattern.Pattern += "{{input}}"
+	}
+
+	// Temporarily replace {{input}} with a sentinel token to protect it
+	// from recursive variable resolution
+	withSentinel := strings.ReplaceAll(pattern.Pattern, "{{input}}", inputSentinel)
+
+	// Process all other template variables in the pattern
+	// At this point, our sentinel ensures {{input}} won't be affected
+	var processed string
+	if processed, err = template.ApplyTemplate(withSentinel, variables, ""); err != nil {
+		return
+	}
+
+	// Finally, replace our sentinel with the actual user input
+	// The input has already been processed for variables if InputHasVars was true
+	pattern.Pattern = strings.ReplaceAll(processed, inputSentinel, input)
+	return
+}
+
+// retrieves a pattern from the database by name
+func (o *PatternsEntity) getFromDB(name string) (ret *Pattern, err error) {
 	patternPath := filepath.Join(o.Dir, name, o.SystemPatternFile)
 
 	var pattern []byte
@@ -29,25 +104,10 @@ func (o *PatternsEntity) Get(name string) (ret *Pattern, err error) {
 	return
 }
 
-// GetApplyVariables finds a pattern by name and returns the pattern as an entry or an error
-func (o *PatternsEntity) GetApplyVariables(name string, variables map[string]string) (ret *Pattern, err error) {
-
-	if ret, err = o.Get(name); err != nil {
-		return
-	}
-
-	if variables != nil && len(variables) > 0 {
-		for variableName, value := range variables {
-			ret.Pattern = strings.ReplaceAll(ret.Pattern, variableName, value)
-		}
-	}
-	return
-}
-
 func (o *PatternsEntity) PrintLatestPatterns(latestNumber int) (err error) {
 	var contents []byte
 	if contents, err = os.ReadFile(o.UniquePatternsFilePath); err != nil {
-		err = fmt.Errorf("could not read unique patterns file. Pleas run --updatepatterns (%s)", err)
+		err = fmt.Errorf("could not read unique patterns file. Please run --updatepatterns (%s)", err)
 		return
 	}
 	uniquePatterns := strings.Split(string(contents), "\n")
@@ -61,8 +121,32 @@ func (o *PatternsEntity) PrintLatestPatterns(latestNumber int) (err error) {
 	return
 }
 
-type Pattern struct {
-	Name        string
-	Description string
-	Pattern     string
+// reads a pattern from a file path and returns it
+func (o *PatternsEntity) getFromFile(pathStr string) (pattern *Pattern, err error) {
+	// Handle home directory expansion
+	if strings.HasPrefix(pathStr, "~/") {
+		var homedir string
+		if homedir, err = os.UserHomeDir(); err != nil {
+			err = fmt.Errorf("could not get home directory: %v", err)
+			return
+		}
+		pathStr = filepath.Join(homedir, pathStr[2:])
+	}
+
+	var content []byte
+	if content, err = os.ReadFile(pathStr); err != nil {
+		err = fmt.Errorf("could not read pattern file %s: %v", pathStr, err)
+		return
+	}
+	pattern = &Pattern{
+		Name:    pathStr,
+		Pattern: string(content),
+	}
+	return
+}
+
+// Get required for Storage interface
+func (o *PatternsEntity) Get(name string) (*Pattern, error) {
+	// Use GetPattern with no variables
+	return o.GetApplyVariables(name, nil, "")
 }
