@@ -1,8 +1,9 @@
 package fsdb
 
 import (
-	goopenai "github.com/sashabaranov/go-openai"
 	"testing"
+
+	goopenai "github.com/sashabaranov/go-openai"
 )
 
 func TestSessions_GetOrCreateSession(t *testing.T) {
